# docs/swagger.yaml
basePath: /
definitions:
...
  calendar.EventDTO:
    properties:
      budgetItemId:
        type: integer
      end:
        type: string
      start:
        type: string
      summary:
        type: string
      uid:
        type: string
    type: object
...
  rest.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: Time planning and tracking API
  title: Klokku API
  version: "0.9"
paths:
...
  /api/calendar/event:
    get:
      description: Retrieve calendar events within a date range
      parameters:
      - description: Start date in RFC3339 format
        in: query
        name: from
        required: true
        type: string
      - description: End date in RFC3339 format
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/calendar.EventDTO'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Get calendar events
      tags:
      - Calendar
    post:
      consumes:
      - application/json
      description: Add a new event to the calendar
      parameters:
      - description: Calendar Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/calendar.EventDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Array of created events (may include recurring instances)
          schema:
            items:
              $ref: '#/definitions/calendar.EventDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Create a calendar event
      tags:
      - Calendar
  /api/calendar/event/{eventUid}:
    delete:
      description: Remove a calendar event by UID
      parameters:
      - description: Event UID
        in: path
        name: eventUid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Delete a calendar event
      tags:
      - Calendar
    put:
      consumes:
      - application/json
      description: Modify an existing calendar event
      parameters:
      - description: Event UID
        in: path
        name: eventUid
        required: true
        type: string
      - description: Updated Calendar Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/calendar.EventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Array of modified events
          schema:
            items:
              $ref: '#/definitions/calendar.EventDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Update a calendar event
      tags:
      - Calendar
  /api/calendar/event/recent:
    get:
      description: Retrieve the most recent calendar events. Specify the number using
        the 'last' query parameter (e.g., last=5)
      parameters:
      - default: 5
        description: Number of recent events to retrieve
        in: query
        name: last
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/calendar.EventDTO'
            type: array
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Get recent calendar events
      tags:
      - Calendar
...
  /api/integrations/clickup/auth:
    delete:
      description: Disconnect and disable the ClickUp integration
      responses:
        "200":
          description: OK
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Disable ClickUp integration
      tags:
      - ClickUp
    get:
      description: Check if the current user has authenticated with ClickUp
      produces:
      - application/json
      responses:
        "200":
          description: "true"
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Not authenticated
      security:
      - XUserId: []
      summary: Check ClickUp authentication status
      tags:
      - ClickUp
  /api/integrations/clickup/auth/callback:
    get:
      description: Handle the OAuth callback from ClickUp
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Redirect to finalUrl with success=true/false
      summary: ClickUp OAuth callback
      tags:
      - ClickUp
  /api/integrations/clickup/auth/login:
    get:
      description: Start the OAuth flow to connect ClickUp account
      parameters:
      - description: URL to redirect to after authentication
        in: query
        name: finalUrl
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OAuth redirect URL
          schema:
            properties:
              redirectUrl:
                type: string
            type: object
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Initiate ClickUp OAuth login
      tags:
      - ClickUp
...
securityDefinitions:
  XUserId:
    description: User ID header required for authentication
    in: header
    name: X-User-Id
    type: apiKey
swagger: "2.0"